Правки после api git tests

+++ ewm-service/src/main/java/ru/yandex/practicum/ewm/service/CompilationServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.yandex.practicum.ewm.dto.compilation.CompilationDto;
 import ru.yandex.practicum.ewm.dto.compilation.NewCompilationDto;
 import ru.yandex.practicum.ewm.dto.compilation.UpdateCompilationRequest;
+import ru.yandex.practicum.ewm.exception.ConflictedDataException;
 import ru.yandex.practicum.ewm.exception.NotFoundException;
 import ru.yandex.practicum.ewm.exception.ValidationException;
 import ru.yandex.practicum.ewm.mapper.CompilationMapper;
@@ -55,7 +56,10 @@ public class CompilationServiceImpl implements CompilationService {
 
     @Override
     public void delCompilation(long compId) {
-        checkCompilation(compId);
+        Compilation compilation = checkCompilation(compId);
+        if (!compilation.getEvents().isEmpty()) {
+            throw new ConflictedDataException("У категории есть связанные события");
+        }
         compilationRepository.deleteById(compId);
     }
